role checks

+++ commands/leagueoflegends/games/delete.js
@@ -6,6 +6,7 @@ const CustomGame = require('../../../database/schema/customgame')
 
 // Functions
 const {
+    checkRole,
     embedConsoleError,
     embedError,
     embedSuccess
@@ -20,18 +21,22 @@ module.exports = class DeleteCustomGameCommand extends Command {
 
     async exec(message) {
         try {
-            const game = await CustomGame.findOne({ guildId: message.guild.id, isActive: true })
+            const permission = checkRole(message, 'Staff')
+
+            if(permission) {
+                const game = await CustomGame.findOne({ guildId: message.guild.id, isActive: true })
     
-            if(game) {
-                message.delete()
-                
-                await game.delete()
-    
-                const embed = embedSuccess(`🕹️ ${message.author}, the active custom game has been deleted!`)
-                return message.channel.send(embed)
-            } else {
-                const embed = embedError(`${message.author}, there was no active custom game found to delete.`)
-                return message.channel.send(embed)
+                if(game) {
+                    message.delete()
+                    
+                    await game.delete()
+        
+                    const embed = embedSuccess(`🕹️ ${message.author}, the active custom game has been deleted!`)
+                    return message.channel.send(embed)
+                } else {
+                    const embed = embedError(`${message.author}, there was no active custom game found to delete.`)
+                    return message.channel.send(embed)
+                }
             }
         } catch(error) {
             const embed = embedConsoleError(error)
